Remove unqualified claim about MongoDB
Unless you plan to explain why MongoDB sucks in the context of security, this statement does nothing to improve the document and instead just makes me suspect that the rest of the document includes similar claims with no backing. In which case, how am I to trust that this guide is dependable or trustworthy?